Moved edit button files
